refactor

+++ src/main/java/com/tamerb/ecommerce/controllers/CategoryController.java
@@ -37,7 +37,7 @@ public class CategoryController {
     }
 
     @PostMapping("/update/{categoryID}")
-    public ResponseEntity<ApiResponse> updateCategory(@PathVariable("categoryID") Integer categoryID, @Valid @RequestBody Category category) {
+    public ResponseEntity<ApiResponse> updateCategory(@PathVariable("categoryID") Long categoryID, @Valid @RequestBody Category category) {
         if(Objects.nonNull(categoryService.readCategory(categoryID))){
             categoryService.updateCategory(categoryID, category);
             return new ResponseEntity<>(new ApiResponse("Category updated successfully", true), HttpStatus.OK);
